se agregan elementos nodos
