+++ src/index.js
@@ -17,13 +17,13 @@ TODO:
 2.- Pasar "root element" a router.
 3.- Invocar el router para renderizar la vista correcta.
 */
-import {Home} from './views/Home.js';
+import {Prueba} from './views/Home.js';
 // ... import other views
 import { setRootEl, setRoutes, onURLChange } from './router.js';
 
 // Define your routes and their associated views
 const routes = {
-  "/": Home,
+  "/": Prueba,
   "/error" : 'Hola',
 };
